Safari-611.1.10/LayoutTests/media/video-extreme-playbackrate-crash-expected.txt


Test that passing extreme playback rate does not cause crash.

EVENT(canplaythrough)
END OF TEST
